+++ Scripts/CustomActions/Build-Source.ps1
@@ -1,2 +1,3 @@
 ﻿# Start the script
-Build-NAVObjects -SetupParameters $SetupParameters -BranchSettings $BranchSettings -TargetFileName Source.txt -IncludeCustomization $false 
+$SaveSourceFileName = Join-Path $SetupParameters.Repository 'Source\Source.txt'
+Build-NAVObjects -SetupParameters $SetupParameters -BranchSettings $BranchSettings -TargetFileName Source.txt -IncludeCustomization $false -SaveSourceFileName $SaveSourceFileName