+++ Scripts/CustomActions/Build-Target.ps1
@@ -1,2 +1,3 @@
 ﻿# Start the script
-Build-NAVObjects -SetupParameters $SetupParameters -BranchSettings $BranchSettings -TargetFileName Target.txt -IncludeCustomization $true
+$SaveSourceFileName = Join-Path $SetupParameters.Repository 'Source\Source.txt'
+Build-NAVObjects -SetupParameters $SetupParameters -BranchSettings $BranchSettings -TargetFileName Target.txt -IncludeCustomization $true -SaveSourceFileName $SaveSourceFileName