+++ Scripts/CustomActions/Load-AppsTools.ps1
@@ -1,0 +1,1 @@
+﻿Load-AppsTools -SetupParameters $SetupParameters