+++ Scripts/CustomActions/Load-InstanceAppTools.ps1
@@ -1,0 +1,1 @@
+﻿Load-InstanceAppTools -SetupParameters $SetupParameters